backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.0

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-genai>=1.53.0
langsmith>=0.3.45

# PDF Generation
reportlab>=4.0.0

# Validation & Utils
pydantic>=2.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
tenacity>=8.2.3
anyio>=4.8.0
orjson>=3.9.14
zstandard>=0.23.0

# Rate Limiting
slowapi>=0.1.9
